pyproject.toml
[project]
name = "job-apply-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "openai>=1.82.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
]
